requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.23
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
PyJWT==2.8.0
python-multipart==0.0.9
pydantic==2.7.0
pydantic-settings==2.1.0
requests==2.31.0
